Prefer current value over accumulator
If the accumulator's score and the current item's score are the same,
use the current item. In the case of a score of 0.0, it is better to get
the current item than the accumulator, because the accumulator value may
be the 'identity' value from `ComputedXor::new`.

It is better to at least get the mask and value from the current value.