Use bytes. Rayon supports `InclusiveRange` now.
Instead of using a Range of integers and casting to byte, just use
bytes. Previously, Rayon did not support `InclusiveRange`, so I had to
convert this to an exclusive Range. However, in order to specify the
upper bound, I need a number one-larger than the max byte. I had to
resort to a Range of integers, which I subsequently cast to byte.

Now, I can specify the inclusive upper bound as a byte (`0xFF`) and just
use bytes everywhere.